[player] improve role player's stability greatly
Prevent the player from crashing when the mplayer subprocess has already
been killed. This happens if you double-click on the skip button or
otherwise are mean to it.

Also made it complain to STDERR if you try to pause or change volume
instead of crashing. STDERR is closed by default, though.