And without further ado: Widgets!
The codebase has been adjusted so that references to non-existant
elements shouldn't break anything, so we can cut out pieces of the UI
and make widgets out of them. This is a simple "Now Playing" widget,
which (through inclusion as an iframe and some javascript) you can set
the player room for, so you can have one for each player. Sample code is
provided in `widget-test.html`, which is set up for my experimental
laptop setup, so you should change the room names.